Sister-pattern size_t casts in multigrid, vhf, np_helper, mcscf
Pattern fixes that match overflow-prone sites we already corrected
elsewhere but missed in their siblings:

* dft/multigrid.c::init_rs_grid: ngrid = mesh[0]*mesh[1]*mesh[2] was
  computed in int and silently under-sized the FFT-grid allocation
  for very fine meshes (~1024 on a side). Cast first factor to size_t.

* vhf/optimizer.c::CVHFset_q_cond / CVHFset_dm_cond: int len parameter
  truncated for nbas >= 46341 (callers pass nbas*nbas), mis-sizing the
  malloc + memcpy. Same class of bug fixed earlier in nr_sr_vhf.c,
  nr_direct.c, fill_nr_s8.c. Promote to size_t.

* np_helper/transpose.c::NPdsymm_021_sum and NPzhermi_021_sum: same
  (size_t)shape[1]*shape[1] cast we applied to NPdtranspose_021 /
  NPztranspose_021, missing from these two siblings.

* mcscf/fci_rdm.c: three pointer-arithmetic sites (ket+stra_id*nb+
  strb_id and bra+stra_id*nb+strb_id, plus the *na variant) missing
  the (size_t) cast already applied at line 83 / 111 in the same file.

+++ pyscf/lib/dft/multigrid.c
@@ -94,7 +94,10 @@ void init_rs_grid(RS_Grid** rs_grid, GridLevel_Info** gridlevel_info, int comp)
     int *mesh = gl_info->mesh;
     rg->data = (double**)malloc(sizeof(double*) * nlevels);
     for (i = 0; i < nlevels; i++) {
-        ngrid = mesh[i*3] * mesh[i*3+1] * mesh[i*3+2];
+        // Cast to size_t before multiplying so very fine meshes
+        // (mesh > ~1024 on a side) do not overflow int and silently
+        // under-size the FFT-grid allocation.
+        ngrid = (size_t)mesh[i*3] * mesh[i*3+1] * mesh[i*3+2];
         (rg->data)[i] = calloc(comp*ngrid, sizeof(double));
     }
     *rs_grid = rg;
