+++ pyscf/lib/np_helper/transpose.c
@@ -132,7 +132,7 @@ void NPdsymm_021_sum(int *shape, double *a, double *out, int hermi)
         shared(shape, a, out, hermi)
 {
         int ic;
-        size_t nn = shape[1] * shape[1];
+        size_t nn = (size_t)shape[1] * shape[1];
 #pragma omp for schedule (static)
         for (ic = 0; ic < shape[0]; ic++) {
                 NPdsymm_sum(shape[1], a+ic*nn, out+ic*nn, hermi);
@@ -146,7 +146,7 @@ void NPzhermi_021_sum(int *shape, double complex *a, double complex *out, int he
         shared(shape, a, out, hermi)
 {
         int ic;
-        size_t nn = shape[1] * shape[1];
+        size_t nn = (size_t)shape[1] * shape[1];
 #pragma omp for schedule (static)
         for (ic = 0; ic < shape[0]; ic++) {
                 NPzhermi_sum(shape[1], a+ic*nn, out+ic*nn, hermi);
